+++ lib/editor.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class EditorView extends StatelessWidget {
+  const EditorView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text("//coming soon"),
+    );
+  }
+}
